loading/lazyload/iframe-loading-eager.tentative.html

<!DOCTYPE html>
<head>
  <title>Iframes with loading='eager' load immediately regardless of their position with respect to the viewport.</title>
  <link rel="author" title="Scott Little" href="mailto:sclittle@chromium.org">
  <link rel="help" href="https://github.com/scott-little/lazyload">
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
</head>

<!--
Marked as tentative until https://github.com/whatwg/html/pull/3752 is landed.
-->

<body>
  <iframe id="in_viewport" src="resources/same-origin-subframe.html?second" loading="eager" width="200px" height="100px">
  </iframe>
  <div style="height:10000px;"></div>

  <!--
    The below_viewport element loads very slowly in order to ensure that the
    window load event is blocked on it.
  -->
  <iframe id="below_viewport" src="slow-redirect.py?location=resources%2Fsame-origin-subframe.html" loading="eager" width="200px" height="100px">
  </iframe>
</body>

<script>
  let t = async_test("Test that iframes with loading='eager' load immediately regardless of their position with respect to the viewport.");

  let has_in_viewport_element_loaded = false;
  let has_in_viewport_subframe_window_loaded = false;
  let in_viewport_frame = document.getElementById("in_viewport");
  in_viewport_frame.testOnSubframeWindowLoad = function() {
    assert_false(has_in_viewport_element_loaded);
    assert_false(has_in_viewport_subframe_window_loaded);
    has_in_viewport_subframe_window_loaded = true;
  };
  in_viewport_frame.addEventListener("load", t.step_func(function() {
    assert_false(has_in_viewport_element_loaded);
    assert_true(has_in_viewport_subframe_window_loaded);
    has_in_viewport_element_loaded = true;
  }));

  let has_below_viewport_element_loaded = false;
  let has_below_viewport_subframe_window_loaded = false;
  let below_viewport_frame = document.getElementById("below_viewport");
  below_viewport_frame.testOnSubframeWindowLoad = function() {
    assert_false(has_below_viewport_element_loaded);
    assert_false(has_below_viewport_subframe_window_loaded);
    has_below_viewport_subframe_window_loaded = true;
  };
  below_viewport_frame.addEventListener("load", t.step_func(function() {
    assert_false(has_below_viewport_element_loaded);
    assert_true(has_below_viewport_subframe_window_loaded);
    has_below_viewport_element_loaded = true;
  }));

  window.addEventListener("load", t.step_func_done(function() {
    assert_true(has_in_viewport_element_loaded);
    assert_true(has_below_viewport_element_loaded);
  }));
</script>
